# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ManuscriptExport": {
        "additionalProperties": false,
        "properties": {
          "branch_id": {
            "pattern": "^[a-z0-9][a-z0-9._:-]{0,127}$",
            "title": "Branch Id",
            "type": "string"
          },
          "checkpoint_id": {
            "anyOf": [
              {
                "pattern": "^[a-z0-9][a-z0-9._:-]{0,127}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkpoint Id"
          },
          "filename": {
            "minLength": 1,
            "title": "Filename",
            "type": "string"
          },
          "markdown": {
            "title": "Markdown",
            "type": "string"
          },
          "project_id": {
            "pattern": "^[a-z0-9][a-z0-9._:-]{0,127}$",
            "title": "Project Id",
            "type": "string"
          }
        },
        "required": [
          "project_id",
          "branch_id",
          "checkpoint_id",
          "filename",
          "markdown"
        ],
        "title": "ManuscriptExport",
        "type": "object"
      },
...
      "ManuscriptReviewOutput": {
        "additionalProperties": false,
        "properties": {
          "review": {
            "$ref": "#/components/schemas/ReviewResult"
          },
          "unsupported_facts": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Unsupported Facts",
            "type": "array"
          }
        },
        "required": [
          "review"
        ],
        "title": "ManuscriptReviewOutput",
        "type": "object"
      },
...
      "NarrativeSourceSummary": {
        "additionalProperties": false,
        "properties": {
          "already_written": {
            "title": "Already Written",
            "type": "boolean"
          },
          "available_viewpoint_ids": {
            "items": {
              "pattern": "^[a-z0-9][a-z0-9._:-]{0,127}$",
              "type": "string"
            },
            "title": "Available Viewpoint Ids",
            "type": "array"
          },
          "boundary": {
            "$ref": "#/components/schemas/SimulationBoundary"
          },
          "branch_id": {
            "pattern": "^[a-z0-9][a-z0-9._:-]{0,127}$",
            "title": "Branch Id",
            "type": "string"
          },
          "checkpoint_id": {
            "pattern": "^[a-z0-9][a-z0-9._:-]{0,127}$",
            "title": "Checkpoint Id",
            "type": "string"
          },
          "event_summary_text": {
            "maxLength": 32768,
            "minLength": 1,
            "title": "Event Summary Text",
            "type": "string"
          },
          "from_step": {
            "minimum": 0.0,
            "title": "From Step",
            "type": "integer"
          },
          "source_id": {
            "pattern": "^[a-z0-9][a-z0-9._:-]{0,127}$",
            "title": "Source Id",
            "type": "string"
          },
          "title_hint": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Title Hint",
            "type": "string"
          },
          "to_step": {
            "minimum": 0.0,
            "title": "To Step",
            "type": "integer"
          }
        },
        "required": [
          "source_id",
          "branch_id",
          "checkpoint_id",
          "from_step",
          "to_step",
          "boundary",
          "title_hint",
          "event_summary_text",
          "available_viewpoint_ids",
          "already_written"
        ],
        "title": "NarrativeSourceSummary",
        "type": "object"
      },
...
      "ReviewIssue": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "evidence_ids": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Evidence Ids",
            "type": "array"
          },
          "message": {
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "severity": {
            "enum": [
              "warning",
              "blocking"
            ],
            "title": "Severity",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "severity"
        ],
        "title": "ReviewIssue",
        "type": "object"
      },
      "ReviewResult": {
        "additionalProperties": false,
        "properties": {
          "issues": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ReviewIssue"
            },
            "title": "Issues",
            "type": "array"
          },
          "mode": {
            "enum": [
              "submission_review",
              "character_review",
              "world_review",
              "promotion_review",
              "manuscript_review"
            ],
            "title": "Mode",
            "type": "string"
          },
          "passed": {
            "title": "Passed",
            "type": "boolean"
          },
          "summary": {
            "minLength": 1,
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "mode",
          "passed",
          "summary"
        ],
        "title": "ReviewResult",
        "type": "object"
      },
      "Scene": {
        "additionalProperties": false,
        "properties": {
          "body": {
            "maxLength": 262144,
            "minLength": 1,
            "title": "Body",
            "type": "string"
          },
          "branch_id": {
            "pattern": "^[a-z0-9][a-z0-9._:-]{0,127}$",
            "title": "Branch Id",
            "type": "string"
          },
          "chapter_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Chapter Id",
            "type": "string"
          },
          "id": {
            "pattern": "^scene-[0-9]{6}$",
            "title": "Id",
            "type": "string"
          },
          "project_id": {
            "pattern": "^[a-z0-9][a-z0-9._:-]{0,127}$",
            "title": "Project Id",
            "type": "string"
          },
          "sequence": {
            "minimum": 1.0,
            "title": "Sequence",
            "type": "integer"
          },
          "source_checkpoint_id": {
            "pattern": "^[a-z0-9][a-z0-9._:-]{0,127}$",
            "title": "Source Checkpoint Id",
            "type": "string"
          },
          "source_event_ids": {
            "items": {
              "pattern": "^[a-z0-9][a-z0-9._:-]{0,127}$",
              "type": "string"
            },
            "minItems": 1,
            "title": "Source Event Ids",
            "type": "array"
          },
          "source_from_step": {
            "minimum": 0.0,
            "title": "Source From Step",
            "type": "integer"
          },
          "source_memory_ids": {
            "items": {
              "pattern": "^[a-z0-9][a-z0-9._:-]{0,127}$",
              "type": "string"
            },
            "title": "Source Memory Ids",
            "type": "array"
          },
          "source_to_step": {
            "minimum": 0.0,
            "title": "Source To Step",
            "type": "integer"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "version": {
            "minimum": 1.0,
            "title": "Version",
            "type": "integer"
          },
          "viewpoint_actor_id": {
            "anyOf": [
              {
                "pattern": "^[a-z0-9][a-z0-9._:-]{0,127}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Viewpoint Actor Id"
          }
        },
        "required": [
          "id",
          "project_id",
          "branch_id",
          "sequence",
          "chapter_id",
          "title",
          "body",
          "source_checkpoint_id",
          "source_from_step",
          "source_to_step",
          "source_event_ids",
          "source_memory_ids",
          "version"
        ],
        "title": "Scene",
        "type": "object"
      },
      "SceneDraft": {
        "additionalProperties": false,
        "properties": {
          "base_scene_version": {
            "default": 0,
            "minimum": 0.0,
            "title": "Base Scene Version",
            "type": "integer"
          },
          "body": {
            "maxLength": 262144,
            "minLength": 1,
            "title": "Body",
            "type": "string"
          },
          "branch_id": {
            "pattern": "^[a-z0-9][a-z0-9._:-]{0,127}$",
            "title": "Branch Id",
            "type": "string"
          },
          "chapter_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Chapter Id",
            "type": "string"
          },
          "id": {
            "pattern": "^scene-[0-9]{6}$",
            "title": "Id",
            "type": "string"
          },
          "project_id": {
            "pattern": "^[a-z0-9][a-z0-9._:-]{0,127}$",
            "title": "Project Id",
            "type": "string"
          },
          "review": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ManuscriptReviewOutput"
              },
              {
                "type": "null"
              }
            ]
          },
          "revision": {
            "default": 0,
            "minimum": 0.0,
            "title": "Revision",
            "type": "integer"
          },
          "sequence": {
            "minimum": 1.0,
            "title": "Sequence",
            "type": "integer"
          },
          "source_checkpoint_id": {
            "pattern": "^[a-z0-9][a-z0-9._:-]{0,127}$",
            "title": "Source Checkpoint Id",
            "type": "string"
          },
          "source_event_ids": {
            "items": {
              "pattern": "^[a-z0-9][a-z0-9._:-]{0,127}$",
              "type": "string"
            },
            "minItems": 1,
            "title": "Source Event Ids",
            "type": "array"
          },
          "source_from_step": {
            "minimum": 0.0,
            "title": "Source From Step",
            "type": "integer"
          },
          "source_memory_ids": {
            "items": {
              "pattern": "^[a-z0-9][a-z0-9._:-]{0,127}$",
              "type": "string"
            },
            "title": "Source Memory Ids",
            "type": "array"
          },
          "source_to_step": {
            "minimum": 0.0,
            "title": "Source To Step",
            "type": "integer"
          },
          "status": {
            "default": "draft",
            "enum": [
              "draft",
              "reviewed",
              "needs_revision",
              "saved"
            ],
            "title": "Status",
            "type": "string"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "viewpoint_actor_id": {
            "anyOf": [
              {
                "pattern": "^[a-z0-9][a-z0-9._:-]{0,127}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Viewpoint Actor Id"
          }
        },
        "required": [
          "id",
          "project_id",
          "branch_id",
          "sequence",
          "chapter_id",
          "title",
          "body",
          "source_checkpoint_id",
          "source_from_step",
          "source_to_step",
          "source_event_ids",
          "source_memory_ids"
        ],
        "title": "SceneDraft",
        "type": "object"
      },
      "SceneGenerationRequest": {
        "additionalProperties": false,
        "properties": {
          "chapter_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Chapter Id",
            "type": "string"
          },
          "checkpoint_id": {
            "pattern": "^[a-z0-9][a-z0-9._:-]{0,127}$",
            "title": "Checkpoint Id",
            "type": "string"
          },
          "from_step": {
            "minimum": 0.0,
            "title": "From Step",
            "type": "integer"
          },
          "to_step": {
            "minimum": 0.0,
            "title": "To Step",
            "type": "integer"
          },
          "viewpoint_actor_id": {
            "anyOf": [
              {
                "pattern": "^[a-z0-9][a-z0-9._:-]{0,127}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Viewpoint Actor Id"
          }
        },
        "required": [
          "checkpoint_id",
          "from_step",
          "to_step",
          "chapter_id"
        ],
        "title": "SceneGenerationRequest",
        "type": "object"
      },
      "SceneMutationResult": {
        "additionalProperties": false,
        "properties": {
          "draft": {
            "$ref": "#/components/schemas/SceneDraft"
          },
          "review": {
            "$ref": "#/components/schemas/ManuscriptReviewOutput"
          },
          "scene": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Scene"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "enum": [
              "saved",
              "rejected"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "draft",
          "review"
        ],
        "title": "SceneMutationResult",
        "type": "object"
      },
      "SceneUpdateRequest": {
        "additionalProperties": false,
        "properties": {
          "body": {
            "maxLength": 262144,
            "minLength": 1,
            "title": "Body",
            "type": "string"
          },
          "expected_revision": {
            "minimum": 0.0,
            "title": "Expected Revision",
            "type": "integer"
          },
          "expected_scene_version": {
            "minimum": 0.0,
            "title": "Expected Scene Version",
            "type": "integer"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "body",
          "expected_revision",
          "expected_scene_version"
        ],
        "title": "SceneUpdateRequest",
        "type": "object"
      },
...
      "SimulationBoundary": {
        "enum": [
          "none",
          "scene",
          "chapter"
        ],
        "title": "SimulationBoundary",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "SessionToken": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Local story-domain sidecar API",
    "title": "AI Story Evolution Engine",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/projects/{project_id}/branches/{branch_id}/manuscript/export": {
      "get": {
        "operationId": "export_manuscript_projects__project_id__branches__branch_id__manuscript_export_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "branch_id",
            "required": true,
            "schema": {
              "title": "Branch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManuscriptExport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionToken": []
          }
        ],
        "summary": "Export Manuscript",
        "tags": [
          "manuscript"
        ]
      }
    },
    "/projects/{project_id}/branches/{branch_id}/manuscript/scenes": {
      "get": {
        "operationId": "list_scenes_projects__project_id__branches__branch_id__manuscript_scenes_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "branch_id",
            "required": true,
            "schema": {
              "title": "Branch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SceneDraft"
                  },
                  "title": "Response List Scenes Projects  Project Id  Branches  Branch Id  Manuscript Scenes Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionToken": []
          }
        ],
        "summary": "List Scenes",
        "tags": [
          "manuscript"
        ]
      }
    },
    "/projects/{project_id}/branches/{branch_id}/manuscript/scenes/generate": {
      "post": {
        "operationId": "generate_scene_projects__project_id__branches__branch_id__manuscript_scenes_generate_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "branch_id",
            "required": true,
            "schema": {
              "title": "Branch Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SceneGenerationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SceneDraft"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionToken": []
          }
        ],
        "summary": "Generate Scene",
        "tags": [
          "manuscript"
        ]
      }
    },
    "/projects/{project_id}/branches/{branch_id}/manuscript/scenes/{scene_id}": {
      "get": {
        "operationId": "get_scene_projects__project_id__branches__branch_id__manuscript_scenes__scene_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "branch_id",
            "required": true,
            "schema": {
              "title": "Branch Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "scene_id",
            "required": true,
            "schema": {
              "title": "Scene Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SceneDraft"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionToken": []
          }
        ],
        "summary": "Get Scene",
        "tags": [
          "manuscript"
        ]
      },
      "put": {
        "operationId": "update_scene_projects__project_id__branches__branch_id__manuscript_scenes__scene_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "branch_id",
            "required": true,
            "schema": {
              "title": "Branch Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "scene_id",
            "required": true,
            "schema": {
              "title": "Scene Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SceneUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SceneMutationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionToken": []
          }
        ],
        "summary": "Update Scene",
        "tags": [
          "manuscript"
        ]
      }
    },
    "/projects/{project_id}/branches/{branch_id}/narrative-sources": {
      "get": {
        "operationId": "list_narrative_sources_projects__project_id__branches__branch_id__narrative_sources_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "branch_id",
            "required": true,
            "schema": {
              "title": "Branch Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after_step",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "After Step"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NarrativeSourceSummary"
                  },
                  "title": "Response List Narrative Sources Projects  Project Id  Branches  Branch Id  Narrative Sources Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "SessionToken": []
          }
        ],
        "summary": "List Narrative Sources",
        "tags": [
          "manuscript"
        ]
      }
    },
...
  }
}